+++ src/test/groovy/ru/vyarus/java/generics/resolver/AssignableTest.groovy
@@ -1,6 +1,7 @@
 package ru.vyarus.java.generics.resolver
 
 import ru.vyarus.java.generics.resolver.context.container.WildcardTypeImpl
+import ru.vyarus.java.generics.resolver.util.TypeLiteral
 import ru.vyarus.java.generics.resolver.util.TypeUtils
 import spock.lang.Specification
 
@@ -39,5 +40,16 @@ class AssignableTest extends Specification {
         TypeUtils.isAssignable(int, Number)
         !TypeUtils.isAssignable(Number, int)
         TypeUtils.isAssignable(Integer, int)
+    }
+
+    def "Check assignability cases"() {
+
+        expect:
+        TypeUtils.isAssignable(left, right) == res
+
+        where:
+        left | right | res
+        new TypeLiteral<Map<String, String>>(){}.getType() | new TypeLiteral<Map<Object, String>>(){}.getType() | true
+        new TypeLiteral<Map<String, String>>(){}.getType() | new TypeLiteral<Map<Integer, String>>(){}.getType() | false
     }
 }